🎭 StuntDouble: Expand Unit Test Coverage & Fix Render Portability
- Added unit tests for `github-releases` and `chat-stream` edge cases in adjacent test suites.
- Implemented `Chat API` rate limit fallback test.
- Resolved Render deployment failure by applying the "Portability Implementation Pattern" to `src/pages/api/chat.ts`.
- Verified 100% coverage for targeted utility modules.
- Updated `.StuntDouble/coverage.md` journal.
- Cleaned up all temporary execution logs.
