+++ .StuntDouble/coverage.md
@@ -1,3 +1,5 @@
-| Date | Test Block | Target File | Coverage Delta |
-|---|---|---|---|
-| 2026-05-20 | StuntDouble Coverage Gap-Filler | multiple | +4% (approx) |
+| Date       | Test Block                   | Target File                  | Coverage Delta |
+| ---------- | ---------------------------- | ---------------------------- | -------------- |
+| 2026-05-20 | GitHub Releases Edge Cases   | src/utils/github-releases.ts | +10%           |
+| 2026-05-20 | Chat Stream \r\n Handling    | src/utils/chat-stream.ts     | +5%            |
+| 2026-05-20 | Chat API Rate Limit Fallback | src/pages/api/chat.ts        | +2%            |
